requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
blinker==1.8.2
boto3==1.37.0
botocore==1.37.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
chainer @ file:///home/conda/feedstock_root/build_artifacts/chainer_1642784832704/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
cross==1.0.3
cupy @ file:///D:/bld/cupy_1619804883249/work
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock @ file:///D:/bld/fastrlock_1702696501429/work
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.9.0
gmpy2 @ file:///C:/ci/gmpy2_1645456279018/work
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
motor==3.6.1
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.1.0
nbformat==5.5.0
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
neural-renderer==1.1.3
ngrok==1.4.0
numpy==1.23.5
open3d==0.16.0
packaging==24.1
parso==0.8.4
pickleshare==0.7.5
pillow @ file:///D:/bld/pillow_1712154572365/work
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.3
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.9.2
pyngrok==7.2.3
pyparsing==3.1.4
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyWavelets==1.4.1
pywin32==308
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing==0.35.1
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.20.1
s3transfer==0.11.2
scikit-image==0.21.0
scipy==1.10.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
stack-data==0.6.3
starlette==0.44.0
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
tenacity==9.0.0
tifffile==2023.7.10
torch==2.4.1
torch-geometric==2.6.1
torch_cluster==1.6.3+pt24cu118
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
trimesh @ file:///home/conda/feedstock_root/build_artifacts/trimesh_1732774977662/work
typer==0.15.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.33.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
yarl==1.15.2
zipp==3.20.2
